+++ objectbox/example/flutter/objectbox_demo_relations/android/app/src/main/kotlin/com/example/objectbox_demo_relations/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.objectbox_demo_relations
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}